fix(BulkDesynth): StdVector indexer takes long, drop the (ulong) cast
Build was failing with "cannot convert from 'ulong' to 'long'". The
loop variable was already `long`; the explicit `(ulong)i` cast was a
copy-paste artifact.

+++ FF14/BulkDesynth/Services/InventoryScanner.cs
@@ -202,7 +202,7 @@ public sealed class InventoryScanner
         var total = sorter->Items.LongCount;
         for (var i = 0L; i < total; i++)
         {
-            var entry = sorter->Items[(ulong)i].Value;
+            var entry = sorter->Items[i].Value;
             if (entry == null) continue;
 
             // entry->Page is an offset 0..3 from InventoryType.Inventory1.
